+++ prisma/schema.prisma
@@ -49,21 +49,21 @@ model Dislike {
 }
 
 model Account {
-    id                  String  @id @default(cuid())
-    userId              String
-    type                String
-    provider            String
-    providerAccountId   String
-    refresh_token       String? @db.Text
-    access_token        String? @db.Text
-    expires_at          Int?
-    token_type          String?
-    scope               String?
-    id_token            String? @db.Text
-    session_state       String?
-    user                User    @relation(fields: [userId], references: [id], onDelete: Cascade)
-    trainedLikeModel    Json?
-    trainedDislikeModel Json?
+    id                        String  @id @default(cuid())
+    userId                    String
+    type                      String
+    provider                  String
+    providerAccountId         String
+    refresh_token             String? @db.Text
+    access_token              String? @db.Text
+    expires_at                Int?
+    token_type                String?
+    scope                     String?
+    id_token                  String? @db.Text
+    session_state             String?
+    user                      User    @relation(fields: [userId], references: [id], onDelete: Cascade)
+    wordFrequencyCountLike    Json?
+    wordFrequencyCountDislike Json?
 
     @@unique([provider, providerAccountId])
     @@index([userId])
